requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
pyodbc==5.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0 